requirements.txt
appnope==0.1.0
auth0-python==2.0.1
backports.shutil-get-terminal-size==1.0.0
click==6.6
configparser==3.5.0
decorator==4.0.10
dotenv==0.0.5
enum34==1.1.6
flake8==3.2.1
Flask==0.11.1
Flask-Cors==3.0.2
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mccabe==0.5.2
pathlib2==2.1.0
pbr==1.8.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycodestyle==2.2.0
pyflakes==1.3.0
Pygments==2.1.3
PyJWT==1.4.2
pymongo==3.4.0
python-http-client==2.2.1
requests==2.12.3
sendgrid==3.6.3
simplegeneric==0.8.1
six==1.10.0
stevedore==1.12.0
traitlets==4.3.1
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wcwidth==0.1.7
Werkzeug==0.11.11
